+++ ChangeLog.md
@@ -21,6 +21,7 @@
 * Ajout du type d'ennemi juré gobelin
 * Implémentation de base de la voie de l'expert du combat
 * La Rune de Puissance du Forgesort maximise tous les dés mais doit s'utiliser avant le lancer des jets de dégâts
+* Implémentation de Contrôle Sanguin (Voie du contrôle corporel, R4)
 
 ### Corrections de bugs
 * Prise en compte de la liberté d'action pour le mot de pouvoir qui immobilise.